hackathon_tutorial/requirements.txt
Flask==3.1.0
gunicorn==23.0.0
python-dotenv==1.1.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
google-generativeai==0.8.5
google-ai-generativelanguage==0.6.15
google-cloud-texttospeech==2.27.0
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
grpcio==1.71.0
grpcio-status==1.71.0
protobuf==5.29.4
requests==2.32.3
Werkzeug==3.1.3
Jinja2==3.1.6
MarkupSafe==3.0.2
click==8.2.0
colorama==0.4.6
itsdangerous==2.2.0
cachetools==5.5.2
httplib2==0.22.0
SQLAlchemy==2.0.40
tqdm==4.67.1
